+++ Cargo.toml
@@ -18,8 +18,9 @@ rand = "~0.6"
 crdts = "4.1.0"
 sha3 = "~0.8.2"
 threshold_crypto = "~0.3.2"
-tiny-keccak = "1.5.0"
-unwrap = "1.2.1"
+tiny-keccak = "~1.5.0"
+unwrap = "~1.2.1"
+crdts = "2.0.0"
 
   [dependencies.serde]
   version = "1.0.91"
