+++ src/lib.rs
@@ -37,6 +37,7 @@ mod mutable_data;
 mod request;
 mod response;
 mod sequence;
+mod transfer;
 mod utils;
 
 pub use errors::{EntryError, Error, Result};
@@ -77,6 +78,7 @@ pub use sequence::{
     UserPermissions as SDataUserPermissions,
 };
 pub use sha3::Sha3_512 as Ed25519Digest;
+pub use transfer::*;
 pub use utils::verify_signature;
 
 use hex_fmt::HexFmt;
@@ -285,94 +287,6 @@ pub enum HandshakeResponse {
     /// Sent by nodes as a response to an invalid `HandshakeRequest::Join` (when a client attempts to join a wrong section).
     InvalidSection,
 }
-
-/// Transfer ID.
-pub type TransferId = u64; // TODO: Use the trait UUID
-
-#[derive(Copy, Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-/// Determines the behaviour of a Transfer
-/// when validation over the different possible states is done.
-pub enum TransferRestrictions {
-    /// Fails transfer if the key has no history.
-    RequireHistory,
-    /// Fails transfer if there are previously recorded transfers.
-    ExpectNoHistory,
-    /// Transfers regardless of previous history.
-    NoRestriction,
-}
-
-///
-#[derive(Copy, Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct Transfer {
-    /// Transfer ID.
-    pub id: TransferId,
-    /// The transfer source.
-    pub from: PublicKey,
-    /// The destination to transfer to.
-    pub to: PublicKey,
-    /// The amount to transfer.
-    pub amount: Money,
-    /// Determines the behaviour of a Transfer.
-    pub restrictions: TransferRestrictions,
-}
-
-/// A Client cmd.
-#[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct ValidateTransfer {
-    /// The transfer.
-    pub transfer: Transfer,
-    /// Client signature over the transfer.
-    pub client_signature: Signature,
-}
-
-/// The Elder event raised when
-/// ValidateTransfer cmd has been successful.
-#[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct TransferValidated {
-    /// The cmd generated by client.
-    pub transfer_cmd: ValidateTransfer,
-    /// Elder signature over the transfer cmd.
-    pub elder_signature: threshold_crypto::SignatureShare,
-    // /// The PK Set of the section
-    // pub pk_set: threshold_crypto::PublicKeySet, // temporary commented out
-}
-
-/// The aggregated Elder signatures of the client transfer cmd.
-#[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct ProofOfAgreement {
-    /// The cmd generated by client.
-    pub transfer_cmd: ValidateTransfer,
-    /// Quorum of Elder sigs over the transfer cmd.
-    pub section_sig: Signature,
-}
-
-/// A Client cmd.
-#[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct RegisterTransfer {
-    /// The transfer proof.
-    pub proof: ProofOfAgreement,
-}
-
-/// The Elder event raised when
-/// RegisterTransfer cmd has been successful.
-#[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-pub struct TransferRegistered {
-    /// The transfer proof.
-    pub proof: ProofOfAgreement,
-}
-
-// /// (Draft) A Client cmd to roll back a failed transfer.
-// #[derive(Clone, Hash, Eq, PartialEq, PartialOrd, Ord, Serialize, Deserialize, Debug)]
-// pub struct CancelTransfer {
-//     /// The transfer id.
-//     pub transfer_id: TransferId,
-//     /// Client signature over the transfer id.
-//     pub client_signature: Signature,
-// }
-
-/// Notification of a Transfer sent to a recipient.
-#[derive(Hash, Eq, PartialEq, PartialOrd, Ord, Clone, Serialize, Deserialize, Debug)]
-pub struct TransferNotification(pub ProofOfAgreement);
 
 #[cfg(test)]
 mod tests {
